backend/requirements.txt
certifi==2021.10.8
django-cors-headers==3.11.0
pytz==2021.3
wincertstore==0.2
